requirements.txt
pandas==2.1.4
openpyxl==3.1.2
snowflake-connector-python==3.12.3
python-dotenv==1.0.0